requirements.txt
django
mysqlclient
requests
python-decouple
Pillow
pytz
django-messages
stravalib
django-unfold